requirement.txt
openai==0.28.1
streamlit
streamlit-extras==0.3.4
langchain==0.0.316
Pillow==9.3.0
PyMuPDF
pandas